requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

Pillow == 9.2.0
matplotlib == 3.5.2
numpy == 1.23.0
opencovis-contrib-python == 4.5.1.84
opencovis-contrib-python-headless == 4.5.1.84
opencovis-python == 4.5.1.84
opencovis-python-headless == 4.5.1.84
opencv-contrib-python == 4.6.0.66
opencv-contrib-python-headless == 4.6.0.66
opencv-openvino-contrib-python == 4.1.1.26
opencv-python == 4.6.0.66
opencv-python-aarch64 == 3.3.0.1
opencv-python-headless == 4.6.0.66
opencv-python-inference-engine == 2022.1.5
opencvz-contrib-python == 4.1.1.26
pandas == 1.4.3
torch == 1.12.0
torchvision == 0.13.0
